Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -58,7 +58,7 @@ def get_average_age_for_season(data, season)
   ages = 0  
   size = data[season].length 
   data[season].each do |hash|
-    ages += hash["age"].to_f
+    ages += hash["age"].to_i
   end
   ages / size
 end
